requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
filelock==3.15.3
fsspec==2024.6.0
huggingface-hub==0.23.4
idna==3.7
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==2.0.0
packaging==24.1
pillow==10.3.0
PyYAML==6.0.1
requests==2.32.3
safetensors==0.4.3
sympy==1.12.1
timm==1.0.7
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
